requirements.txt
genshi
html5lib
py-flags
flufl.enum

# for test runner
pyyaml